init models

+++ src/main/java/com/main/ecommerce/repository/ProductRepository.java
@@ -1,10 +1,10 @@
 package com.main.ecommerce.repository;
 
-import com.main.ecommerce.model.Product;
+import com.main.ecommerce.model.products.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ProductRepository extends JpaRepository<Product, Integer> {
+public interface ProductRepository extends JpaRepository<Product, Long> {
 
 }
